Jinja environments create on the fly.
This commit solves a problem running Django tests. Some of the Django
tests alter TEMPLATE_DIRS, and if that test runs first, it remains
stuck in the Jinja environment.

Creating the environment on the fly each time means that the settings
are checked anew each time a JinjaTemplateResponse is created.